# openAPI/openAPI-spec.yaml
openapi: 3.0.0
info:
  title: Chatbot API with Google Authentication
  description: An API for interacting with our chatbot, providing feedback, and managing sessions.
  version: 1.0.2
externalDocs:
  description: Find more info here
  url: https://example.com
servers:
  - url: https://api.yourdomain.com/v1
paths:
  /login:
    get:
      summary: Redirect to Google's OAuth 2.0 authentication
      responses:
        '302':
          description: Redirect to Google's authentication page
        '400':
          description: Bad request - Malformed request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Authentication failed
        '403':
          description: Forbidden - Access denied
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/callback:
    get:
      summary: Callback endpoint for Google's OAuth 2.0 authentication
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: Authorization code from Google.
      responses:
        '200':
          description: Successful authentication and user information returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfoResponse'
        '401':
          description: Unauthorized - Authentication failed or non-BU email used
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chat:
    post:
      summary: Send a question to the chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequestWithSession'
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: JWT token for authentication.
      responses:
        '202':
          description: Accepted but processing incomplete. More information needed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncompleteResponse'
        '400':
          description: Bad request - Malformed request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Missing, invalid token, or session expired
        '429':
          description: Too many requests
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /chat/{responseID}/result:
    get:
      summary: Poll for the chatbot's response using the responseID
      parameters:
        - in: path
          name: responseID
          schema:
            type: string
          required: true
          description: The unique ID of the chatbot's initial acknowledgment.
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '202':
          description: Processing still ongoing
        '401':
          description: Unauthorized - Missing or invalid token/session
        '404':
          description: ResponseID not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /feedback:
    post:
      summary: Provide feedback on the chatbot's response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackRequest'
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: JWT token for authentication.
      responses:
        '200':
          description: Feedback recorded successfully
        '400':
          description: Bad request - Malformed request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Missing or invalid token/session
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    ChatRequestWithSession:
      type: object
      properties:
        sessionID:
          type: string
          description: Identifier for the user's current session.
        question:
          type: string
          description: The question posed by the user.
    ChatResponse:
      type: object
      properties:
        response:
          type: string
          description: The chatbot's response.
        responseID:
          type: string
          description: A unique ID for the response.
    FeedbackRequest:
      type: object
      properties:
        responseID:
          type: string
          description: The ID of the response being rated.
        liked:
          type: boolean
          description: True if the user liked the response, false otherwise.
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: A descriptive error message.
        code:
          type: integer
          description: A specific error code for more detailed debugging.
    IncompleteResponse:
      type: object
      properties:
        message:
          type: string
          description: Information or clarification needed from the user.
    UserInfoResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token for the user.
        email:
          type: string
          description: User's email address.
        first_name:
          type: string
          description: User's first name.
        last_name:
          type: string
          description: User's last name.
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          scopes:
            openid: Google login
            profile: Google profile information
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - GoogleOAuth: []
  - BearerToken: []